requirements.txt
langchain
langchain-community
openai
faiss-cpu
pandas
python-dotenv
streamlit